+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod board;
-pub mod solver;
+pub mod mate;
 pub mod wasm;
 
 pub use wasm::{decode_x, decode_y, encode_xy, solve_vcf};
